Split coding-plan module and drop dead fields
coding_plan.rs crossed the 1000-line bar with three responsibilities; split HTML scrape into coding_plan_html.rs. Remove RemainsRow.usage/limit + boost-permille quota plumbing computed but never consumed.
